+++ packages/slack/src/index.ts
@@ -1,7 +1,7 @@
-import { connector } from './connector'
-import { serveIfEntryPoint } from './entry'
+import { connector } from "./connector";
+import { serveIfEntryPoint } from "./entry";
 
-export { connector }
-export default connector
+export { connector };
+export default connector;
 
-await serveIfEntryPoint(import.meta.url)
+await serveIfEntryPoint(import.meta.url);